+++ tools/extract-po-files.py
@@ -134,7 +134,7 @@ else:
         updated_langs.sort()
         if len(updated_langs) > 1:
             updated_langs_english = ', '.join(updated_langs[:-1])
-            updated_langs_english = updated_langs_english[2:] + ' and %s' % updated_langs[-1]
+            updated_langs_english = updated_langs_english + ' and %s' % updated_langs[-1]
             print('Updated {} translations.\n'.format(updated_langs_english))
         else:
             print('Updated {} translation.\n'.format(updated_langs[0]))